Add files via upload
This capstone project is a "structured walkthrough" penetration test of a fictional company, Artemis, Incorporated (“Artemis”).  In this "structured walkthrough", a group of peers reviews and discusses the technical aspects of various IT, IT Security, and IT Audit work products to find errors and to improve the quality of the final product or service. This project was 5 phases in length. 

Phase 1: Perform Reconnaissance
A 2-3 page description of all the tools and methods I would use to perform reconnaissance.

Phase 2: Identify Targets and Run Scans
A 2-3 page description of the tools I would plan on using for the network scans, explaining my reasoning for selecting them, and how I would use them.

Phase 3: Identify Vulnerabilities
A 2-3 page description of tools I would use for the vulnerability scans, how I would use them, screenshots of the tools with configuration options and settings, and the pros and cons of each tool.

Phase 4: Threat Assessment
A 10 page document showing the work I did on threat assessment.

Phase 5: Reporting
Two mock reports; An Executive Summary for a client’s senior management, and a Detailed Technical Report for the client’s IT staff based on the scenario this capstone project was based on.